Struct rollbar::ReportErrorBuilder

pub struct ReportErrorBuilder<'a> { /* fields omitted */ }

Builder specialized for reporting errors.

Methods

impl<'a> ReportErrorBuilder<'a>

fn with_backtrace(&'a mut self, backtrace: &'a Backtrace) -> &'a mut Self

Attach a backtrace::Backtrace to the description of the report.

fn with_frame(&'a mut self, frame_builder: FrameBuilder<'a>) -> &'a mut Self

Add a new frame to the collection of stack frames.

fn with_level<T>(&'a mut self, level: T) -> &'a mut Self where T: Into<Level>

Set the security level of the report. Level::ERROR is the default value.

fn with_title<T>(&'a mut self, title: T) -> &'a mut Self where T: Into<String>

Set the title to show in the dashboard for this report.

fn send(&mut self) -> JoinHandle<Option<ResponseStatus>>

Send the report to Rollbar.